% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaHatEff.R
\name{alphaHatEff}
\alias{alphaHatEff}
\title{alphaHatEff}
\usage{
alphaHatEff(Y, X, Z, Qn1, Qn0, gn, select)
}
\arguments{
\item{Y}{A matrix with columns corresponding to different outcomes in training data}

\item{Qn1}{A matrix of predictions from SuperLearner on training data with Z = 1}

\item{Qn0}{A matrix of predictions from SuperLearner on training data with Z = 0}

\item{gn}{A matrix of propensity estimates in training data}

\item{QnObs}{A matrix of predictions from SuperLearner on training data with observed Z}
}
\value{
weights A numeric vector of weights of the same length as the number
of outcomes considered. The weights sum to 1.
}
\description{
Compute weights for ate estimation
}

