% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeR2.R
\name{computeR2}
\alias{computeR2}
\title{Compute a weighted R-squared from CV.SuperLearner objects.}
\usage{
computeR2(Ymat, folds, pred, validPred, weightList, V, J, n)
}
\arguments{
\item{Ymat}{A matrix of outcomes on the the full data set with columns 
corresopnding to different outcomes.}

\item{folds}{A list of the same length as V that was used in the 
cross-validation procedure. Each object of the list is a vector specifying
the indices of the rows that made up the validation set for that split.}

\item{pred}{A list of the same length as V that was used in the 
cross-validation procedure containing the predictions made on the training 
sample by the prediction algorithms. Each entry should be a matrix with a 
column for a prediction of each outcome.}

\item{validPred}{A list the same length as V that was used in the 
cross-validation procedure containing the predictions made on the validation 
sample by the prediction algorithms. Each entry should be matrix with a 
column for a prediction of each outcome.}

\item{weightList}{A list the same length as V that was used in the 
cross-validation procedure containing the predictions made on the validation 
sample by the prediction algorithms. Should be matrix with a column for a 
prediction of each outcome.}

\item{V}{The number of sample paritions used in cross-validation procedure.}

\item{J}{The number of different outcomes.}

\item{n}{The number of observations.}
}
\value{
cv.wR2 An estimated of the weighted R-squared

cv.wR2.ci A 95\% confidence interval for the weighted R-squared

cv.wSummary A data.frame illustrating the weights and R-squared 
estimated in each fold

icList A list of estimated influence functions

numList A list of estimated cross-validated weighted MSE (numerator 
of R-squared)

denomList A list of estimated variances of weighted outcome 
(denominator of R-squared)

n The sample size

V The number of partitions

foldnList A list of the number in each partition
}
\description{
This function takes a matrix of various outcomes and cross-validated 
precitions of these outcomes, along with a set of weights for the outcomes, 
and comptues the R-squared for predicting the specified weighted outcome 
using (the same) specified weighted combination of predictions.
}

