% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPredictionsOnValidation.R
\name{getPredictionsOnValidation}
\alias{getPredictionsOnValidation}
\title{getPredictionsOnValidation}
\usage{
getPredictionsOnValidation(out, whichAlgorithm)
}
\arguments{
\item{out}{List of \code{CV.SuperLearner} objects}

\item{whichAlgorithm}{A \code{character} of which algorithm to obtain validation predictions for.}
}
\description{
Get validation predictions from a CV.SuperLearner object.
}

