% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{getTrueUnivarateR2}
\alias{getTrueUnivarateR2}
\title{getTrueUnivariateR2}
\usage{
getTrueUnivarateR2(n = 1e+06)
}
\description{
Compute the true value of R^2_{0,j}. The code computes it only for 
y1 (from \code{makeData}), but it will be the same for all outcomes
}

