% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{getTrueUnivariateR2}
\alias{getTrueUnivariateR2}
\title{getTrueUnivariateR2}
\usage{
getTrueUnivariateR2(n = 1e+06)
}
\arguments{
\item{n}{Sample size}
}
\description{
Compute the true value of R^2_{0,j}. The code computes it only for 
y1 (from \code{makeData}), but it will be the same for all outcomes
}

