% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{getTrueUnivariateR2Exclude}
\alias{getTrueUnivariateR2Exclude}
\title{getTrueUnivariateR2Exclude}
\usage{
getTrueUnivariateR2Exclude(n = 1e+06, excludeX, whichY)
}
\arguments{
\item{n}{Sample size}

\item{excludeX}{What X column to exclude}

\item{whichY}{What Y column to study}
}
\description{
Compute the true value of R^2_{0,j} when a given variable of X is excluded.
}

