% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{getTrueWeights}
\alias{getTrueWeights}
\title{getTrueWeights}
\usage{
getTrueWeights(n = 1e+06)
}
\arguments{
\item{n}{Sample size}
}
\description{
Compute the true value of the optimal weights for simulation 1.
}
