% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnivariateR2.R
\name{getUnivariateR2}
\alias{getUnivariateR2}
\title{getUnivariateR2}
\usage{
getUnivariateR2(Y, psiHat.Pnv0, return.IC)
}
\arguments{
\item{Y}{The matrix of outcomes}

\item{psiHat.Pnv0}{The predictions on training sample}

\item{return.IC}{Boolean for whether to return influence functions}
}
\value{
A \code{list} with point estimate and 95% Wald-style confidence interval 
for R^2 for each outcome individually.
}
\description{
Comptues the R^2 and a confidence interval for a particular outcome and 
Super Learner prediction.
}

