% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{plotSimResults}
\alias{plotSimResults}
\title{plotSimResults}
\usage{
plotSimResults(out, outExclude, ...,
  saveDir = "~/Dropbox/Job Applications/Job Talk/",
  outFile = "sim1Rslt.pdf", outExFile = "sim2Rslt.pdf")
}
\arguments{
\item{out}{A \code{data.frame} of output that results from running \code{sce.sh}}

\item{outExclude}{A \code{data.frame} of output that results from running \code{sce2.sh}}

\item{...}{Passed to \code{pdf}}

\item{saveDir}{A \code{character} specifying the save directory}

\item{outFile}{A \code{character} specifying the .pdf file name for simulation 1 results}

\item{outExFile}{A \code{character} specifying the .pdf file name for simulation 2 results}
}
\description{
Function to plot bias and coverage from simulations
}
