% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optWeight-print-method.R
\name{print.optWeight}
\alias{print.optWeight}
\title{print.optWeight}
\usage{
\method{print}{optWeight}(x, ...)
}
\arguments{
\item{x}{An object of class \code{optWeight}}

\item{...}{Other args (not used)}
}
\description{
Print output of optimal weight procedure
}
