% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optWeight-print-method.R
\name{print.r2.optWeight}
\alias{print.r2.optWeight}
\title{print.r2.optWeight}
\usage{
\method{print}{r2.optWeight}(object, ...)
}
\description{
Print the output of R-squared for Super Learner of optimal R-squared
}

