% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2Diff-print-method.R
\name{print.r2_diff}
\alias{print.r2_diff}
\title{print.r2_diff}
\usage{
\method{print}{r2_diff}(x, ...)
}
\arguments{
\item{x}{An object of class \code{r2_diff}}

\item{...}{Other arguments (passed to \code{round})}
}
\description{
Print output of optimal weight procedure
}

