% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.r2_optWeight}
\alias{print.r2_optWeight}
\title{print.r2_optWeight}
\usage{
\method{print}{r2_optWeight}(x, ...)
}
\arguments{
\item{x}{An object of class r2_optWeight}

\item{...}{Other args (not used)}
}
\description{
Print the output of R-squared for Super Learner of optimal R-squared
}
