% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.optWeight.R
\name{r2.optWeight}
\alias{r2.optWeight}
\title{r2.optWeight}
\usage{
r2.optWeight(object, Y, X, evalV = 20, return.IC = TRUE, seed = 12345,
  ...)
}
\arguments{
\item{object}{A \code{optWeight} object.}

\item{Y}{The \code{data.frame} of outcomes that was used to fit \code{object}}

\item{X}{The \code{data.frame} of predictors that was used to fit \code{object}}

\item{evalV}{The number of outer cross validation folds to use to evaluate the predictive 
performance of \code{object} for predicting the optimal combined outcome.}

\item{return.IC}{A \code{boolean} indicating whether to return estimated influence
function at the observed data values (needed for post-hoc comparisons).}
}
\value{
An cross-validated estimate of the R-squared for the optimal prediction and 
standard error and confidence interval.
}
\description{
Estimate the performance of the Super Learner for predicting the optimal weighted 
combination via cross-validation.
}
\examples{
# Example 1 -- simple fit 
X <- data.frame(x1=runif(n=100,0,5), x2=runif(n=100,0,5))
Y1 <- rnorm(100, X$x1 + X$x2, 1)
Y2 <- rnorm(100, X$x1 + X$x2, 3)
Y <- data.frame(Y1 = Y1, Y2 = Y2)
fit <- optWeight(Y = Y, X = X, SL.library = c("SL.glm","SL.mean"), family = gaussian(),outerV = 10, return.CV.SuperLearner = FALSE)
perf.fit <- r2.optWeight(object = fit, Y = Y, X = X, evalV = 5)

}

