% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.optWeight.R
\name{r2_optWeight}
\alias{r2_optWeight}
\title{r2_optWeight}
\usage{
r2_optWeight(object, Y, X, evalV = 10, return.IC = TRUE, seed = 12345,
  verbose = FALSE, parallel = FALSE, n.cores = parallel::detectCores(),
  ...)
}
\arguments{
\item{object}{A \code{optWeight} object}

\item{Y}{The \code{data.frame} of outcomes that was used to fit \code{object}}

\item{X}{The \code{data.frame} of predictors that was used to fit \code{object}}

\item{evalV}{The number of outer cross validation folds to use to evaluate the predictive 
performance of \code{object} for predicting the optimal combined outcome}

\item{return.IC}{A \code{boolean} indicating whether to return estimated influence
function at the observed data values (needed for post-hoc comparisons)}

\item{seed}{Random seed to set}

\item{verbose}{A \code{boolean} indicating whether to show a progress bar}

\item{parallel}{A \code{boolean} indicating whether to run the CV.SuperLearner calls
in parallel using \code{mclapply}. Be sure to set options()$mc.cores to}

\item{n.cores}{A \code{numeric} indicating how many cores to use if \code{parallel = TRUE} 
By default will use \code{parallel::detectCores()}}

\item{...}{Other args (not currently used)}
}
\value{
An cross-validated estimate of the R-squared for the optimal prediction and 
standard error and confidence interval.
}
\description{
Estimate the performance of the Super Learner for predicting the optimal weighted 
combination via cross-validation.
}
\examples{
# Example 1 -- simple fit 
set.seed(1234)
X <- data.frame(x1=runif(n=100,0,5), x2=runif(n=100,0,5))
Y1 <- rnorm(100, X$x1 + X$x2, 1)
Y2 <- rnorm(100, X$x1 + X$x2, 3)
Y <- data.frame(Y1 = Y1, Y2 = Y2)
fit <- optWeight(Y = Y, X = X, SL.library = c("SL.glm","SL.mean"), 
family = "gaussian",outerV = 10, 
return.CV.SuperLearner = FALSE)
perf.fit <- r2_optWeight(object = fit, Y = Y, X = X, evalV = 5)

}

