% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2weight.R
\name{r2weight}
\alias{r2weight}
\title{Compute the optimally weight R-squared from CV.SuperLearner objects.}
\usage{
r2weight(cvslList, X, whichAlgorithm = "SuperLearner")
}
\arguments{
\item{cvslList}{A list of CV.SuperLearner objects on different outcomes 
that were executed using an identical sample partitioning with 
option saveFitLibrary=TRUE.  \code{cvslList}}

\item{X}{A list of data.frames that were used in CV.SuperLearner fits. The 
length of X should be either one (if same data.frame was used for all 
CV.SuperLearner fits or should be equal to length(cvslList). \code{X}}

\item{whichAlgorithm}{A character indicating which algorithm from the 
SuperLearner library for which compute cross-validated R-squared should be 
computed. The default is "SuperLearner"; other options are 
"discreteSuperLearner" or the name of a SuperLearner wrapper function used 
in the call to CV.SuperLearner.}
}
\value{
r2weight A list containing weighted R-squared output

r2 A list containing single outcome R-squared output
}
\description{
This function takes a list of CV.SuperLearner objects that were fit using 
an identical sample partitioning on different outcomes. The function computes
the R-squared for the optimal weighted linear combination of the outcomes and 
associated confidence intervals. The function also returns the estimated 
R-squared for each outcome individually.
}
\examples{
library(SuperLearner)
library(Rsolnp)
X <- data.frame(x1=runif(n=100,0,5), x2=runif(n=100,0,5))
Y1 <- rnorm(100, X$x1 + X$x2, 1)
Y2 <- rnorm(100, X$x1 + X$x2, 3)

# set seed to ensure identical sample partition
set.seed(1234)
cvsl1 <- CV.SuperLearner(Y=Y1, X=X, SL.library=c("SL.glm","SL.step"),
control=list(saveFitLibrary=TRUE))
set.seed(1234)
cvsl2 <- CV.SuperLearner(Y=Y2, X=X, SL.library=c("SL.glm","SL.step"),
control=list(saveFitLibrary=TRUE))

out <- r2weight(cvslList=list(cvsl1,cvsl2), X=list(X))
out
}
\keyword{R-squared,}
\keyword{cross-validation,}
\keyword{prediction}

