% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.varImport-summary-method.R
\name{summary.r2.varImport}
\alias{summary.r2.varImport}
\title{summary.r2.varImport}
\usage{
\method{summary}{r2.varImport}(object, ...)
}
\arguments{
\item{object}{An object of class \code{r2.varImport}}
}
\description{
Method for summarizing output of r2.varImport.
}

