% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2VarImp-summary-method.R
\name{summary.r2_varImp}
\alias{summary.r2_varImp}
\title{summary.r2_varImp}
\usage{
\method{summary}{r2_varImp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{r2.varImport}}

\item{...}{Other arguments}
}
\description{
Method for summarizing output of r2.varImport.
}

