% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunctions.R
\name{getTrueOptR2Exclude}
\alias{getTrueOptR2Exclude}
\title{getTrueOptR2Exclude}
\usage{
getTrueOptR2Exclude(n = 1e+06, excludeX)
}
\arguments{
\item{n}{Sample size}

\item{excludeX}{What X column to drop}
}
\description{
Compute the true value of R^2_{0,omega_0} when a given variable of X is excluded.
}
