% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict.optWeight}
\alias{predict.optWeight}
\title{predict.optWeight}
\usage{
predict.optWeight(object, newdata, ...)
}
\arguments{
\item{object}{A fit \code{optWeight} object.}

\item{newdata}{A new data frame with the same design as the original \code{X} object
that was used in the call to \code{optWeight}.}

\item{...}{Other args (not used)}
}
\description{
A method for prediction of weighted outcome based on \code{optWeight} fitted object.
}
